--- modules/app/constants.ts
@@ -1,1 +1,0 @@
-export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
